+++ src/editors/meshex.rs
@@ -1,8 +1,9 @@
 use crate::{path::folder_editor_title, EditorResponse};
-use egui::{special_emojis::GITHUB, Grid, ScrollArea};
+use egui::{special_emojis::GITHUB, Grid, Label, Response, ScrollArea, Sense, Ui};
 use log::error;
 use rfd::FileDialog;
 use ssbh_data::prelude::*;
+use ssbh_wgpu::RenderModel;
 use std::path::Path;
 
 pub fn meshex_editor(
@@ -11,6 +12,7 @@ pub fn meshex_editor(
     file_name: &str,
     meshex: &mut MeshExData,
     mesh: Option<&MeshData>,
+    render_model: &mut Option<&mut RenderModel>,
 ) -> EditorResponse {
     let mut open = true;
     let mut changed = false;
@@ -88,9 +90,23 @@ pub fn meshex_editor(
                         ui.end_row();
 
                         for group in &mut meshex.mesh_object_groups {
-                            for entry in &mut group.entry_flags {
-                                ui.label(&group.mesh_object_full_name);
-                                ui.label(&group.mesh_object_name);
+                            for (subindex, entry) in group.entry_flags.iter_mut().enumerate() {
+                                // Avoid short-circuiting below to always show both labels.
+                                let response1 = hoverable_label(ui, &group.mesh_object_full_name);
+                                let response2 = hoverable_label(ui, &group.mesh_object_name);
+
+                                if response1.hovered() || response2.hovered() {
+                                    if let Some(render_mesh) = render_model.as_mut().and_then(|m| {
+                                        m.meshes.iter_mut().find(|m| {
+                                            m.name == group.mesh_object_full_name
+                                                && m.subindex == subindex as u64
+                                        })
+                                    }) {
+                                        // Outline the selected mesh in the viewport.
+                                        render_mesh.is_selected = true
+                                    }
+                                }
+
                                 changed |=
                                     ui.checkbox(&mut entry.draw_model, "Draw Model").changed();
                                 changed |=
@@ -107,4 +123,8 @@ pub fn meshex_editor(
         changed,
         saved,
     }
+}
+
+fn hoverable_label(ui: &mut Ui, label: &str) -> Response {
+    ui.add(Label::new(label).sense(Sense::click()))
 }